Update settings transition
