+++ app/src/main/java/com/capyreader/app/ui/settings/SettingsPanelScaffold.kt
@@ -1,11 +1,7 @@
 package com.capyreader.app.ui.settings
 
-import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.ArrowBack
 import androidx.compose.material.icons.rounded.Close
@@ -18,15 +14,12 @@ import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults.exitUntilCollapsedScrollBehavior
-import androidx.compose.material3.TopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import com.capyreader.app.ui.isAtMostMedium
 import com.capyreader.app.ui.isExpanded
 import com.capyreader.app.ui.settings.panels.SettingsPanel
@@ -38,15 +31,7 @@ fun SettingsPanelScaffold(
     onBack: () -> Unit,
     content: @Composable () -> Unit,
 ) {
-    val topBarState = rememberSaveable(panel, saver = TopAppBarState.Saver) {
-        TopAppBarState(
-            initialHeightOffsetLimit = 0f,
-            initialHeightOffset = 0f,
-            initialContentOffset = 0f,
-        )
-    }
-
-    val scrollBehavior = exitUntilCollapsedScrollBehavior(state = topBarState)
+    val scrollBehavior = exitUntilCollapsedScrollBehavior()
     val snackbarHost = remember { SnackbarHostState() }
 
     CompositionLocalProvider(
